# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Acontext.",
        "title": "Acontext API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/project": {
            "post": {
                "security": [
                    {
                        "RootAuth": []
                    }
                ],
                "description": "Create a new project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Create project",
                "parameters": [
                    {
                        "description": "CreateProject payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateProjectReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Project"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session": {
            "post": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Create a new session under a space",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Create session",
                "parameters": [
                    {
                        "description": "CreateSession payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSessionReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Session"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/{session_id}": {
            "delete": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Delete a session by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Delete session",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/session/{session_id}/configs": {
            "get": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Get session configs by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Get session configs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Session"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Update session configs by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Update session configs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateSessionConfigs payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateSessionConfigsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/session/{session_id}/connect_to_space": {
            "post": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Connect a session to a space by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Connect session to space",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConnectToSpace payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ConnectToSpaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/session/{session_id}/messages": {
            "post": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Supports JSON and multipart/form-data. In multipart mode: the payload is a JSON string placed in a form field.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Send message to session",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SendMessage payload (Content-Type: application/json)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SendMessageReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "SendMessage payload (Content-Type: multipart/form-data)",
                        "name": "payload",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "When uploading files, the field name must correspond to parts[*].file_field.",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ConnectToSpaceReq": {
            "type": "object",
            "required": [
                "space_id"
            ],
            "properties": {
                "space_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "handler.CreateProjectReq": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "handler.CreateSessionReq": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                },
                "space_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-42661417"
                }
            }
        },
...
        "handler.SendMessageReq": {
            "type": "object",
            "required": [
                "parts",
                "role"
            ],
            "properties": {
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.PartIn"
                    }
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "handler.UpdateSessionConfigsReq": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "model.Asset": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "number"
                },
                "etag": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "description": "Asset \u003c-\u003e Message",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "mime": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "size_b": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "assets": {
                    "description": "Message \u003c-\u003e Asset",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Asset"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_assets": {
                    "description": "Message \u003c-\u003e MessageAsset",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MessageAsset"
                    }
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "role": {
                    "type": "string"
                },
                "session": {
                    "description": "Message \u003c-\u003e Session",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Session"
                        }
                    ]
                },
                "session_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.MessageAsset": {
            "type": "object",
            "properties": {
                "asset": {
                    "description": "MessageAsset \u003c-\u003e Asset",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Asset"
                        }
                    ]
                },
                "asset_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "description": "MessageAsset \u003c-\u003e Message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Message"
                        }
                    ]
                },
                "message_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Project": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "sessions": {
                    "description": "Project \u003c-\u003e Session",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Session"
                    }
                },
                "spaces": {
                    "description": "Project \u003c-\u003e Space",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Session": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "description": "Session \u003c-\u003e Message",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "project": {
                    "description": "Session \u003c-\u003e Project",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Project"
                        }
                    ]
                },
                "project_id": {
                    "type": "string"
                },
                "space": {
                    "description": "Session \u003c-\u003e Space",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Space"
                        }
                    ]
                },
                "space_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Space": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project": {
                    "description": "Space \u003c-\u003e Project",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Project"
                        }
                    ]
                },
                "project_id": {
                    "type": "string"
                },
                "sessions": {
                    "description": "Space \u003c-\u003e Session",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Session"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.PartIn": {
            "type": "object",
            "properties": {
                "file_field": {
                    "description": "File field name in the form",
                    "type": "string"
                },
                "meta": {
                    "description": "[Optional] metadata",
                    "type": "object",
                    "additionalProperties": true
                },
                "text": {
                    "description": "Text sharding",
                    "type": "string"
                },
                "type": {
                    "description": "\"text\" | \"image\" | ...",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ProjectAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RootAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}